requirements.txt
tqdm
wheel
jaconv
rdflib
mysql-connector-python
pandas
networkx
matplotlib
openpyxl